+++ CHANGELOG.rst
@@ -1,7 +1,8 @@
 next (tbd)
 -------------------
 
-* `#16 <https://github.com/pytest-dev/pytest-reportlog/issues/15>`_: switch recommended file extension to jsonj
+* `#16 <https://github.com/pytest-dev/pytest-reportlog/issues/16>`_: switch recommended file extension to jsonl
+* `#17 <https://github.com/pytest-dev/pytest-reportlog/issues/17>`_: add compression support for .bz2/.gz/.xz log file extensions
 
 
 0.3.0 (2023-04-26)