+++ CHANGELOG.rst
@@ -1,5 +1,5 @@
-Next (tbd)
--------------------
+0.4.0 (2023-05-22)
+------------------
 
 * `#15 <https://github.com/pytest-dev/pytest-reportlog/issues/15>`_: switch recommended file extension to `jsonl <https://jsonlines.org/>`__.
 * `#17 <https://github.com/pytest-dev/pytest-reportlog/issues/17>`_: add compression support for ``.bz2``/``.gz``/``.xz`` log file extensions.